Version change with new MIT License
For greater portability of the software with systems such as MAGMA we, the authors Peter A. Brooksbank, Joshua F. Maglione, and James B. Wilson,  have revisited the GPL License and chosen to use the MIT license going forward.  Any forks and branches from this commit onward are under this license.